Fix Robot, Vivid and.... Add Flat?
Robot and Vivid got added to the ITGmania Beta so I tidied those up in various ways. All of which I had already done for myself on my ITGmania before the implementation news to try and remake ITG's stuff so it was just a matter of getting the files sorted out here.